Crate aws_sdk_sso

AWS Single Sign-On Portal is a web service that makes it easy for you to assign user access to AWS SSO resources such as the user portal. Users can get AWS account applications and roles assigned to them and get federated into the application.

For general information about AWS SSO, see What is AWS Single Sign-On? in the AWS SSO User Guide.

This API reference guide describes the AWS SSO Portal operations that you can call programatically and includes detailed information on data types and errors.

AWS provides SDKs that consist of libraries and sample code for various programming languages and platforms, such as Java, Ruby, .Net, iOS, or Android. The SDKs provide a convenient way to create programmatic access to AWS SSO and other AWS services. For more information about the AWS SDKs, including how to download and install them, see Tools for Amazon Web Services.

Crate Organization

The entry point for most customers will be Client. Client exposes one method for each API offered by the service.

Some APIs require complex or nested arguments. These exist in model.

Lastly, errors that can be returned by the service are contained within error. Error defines a meta error encompassing all possible errors that can be returned by the service.

The other modules within this crate are not required for normal usage.

Modules

client

Client and fluent builders for calling the service.

config

Configuration for the service.

error

Errors that can occur when calling the service.

input

Input structures for operations.

lens

Generated accessors for nested fields

middleware

Base Middleware Stack

model

Data structures used by operation inputs/outputs.

operation

All operations that this crate can perform.

output

Output structures for operations.

paginator

Paginators for the service

types

Re-exported types from supporting crates.

Structs

AppName

App name that can be configured with an AWS SDK client to become part of the user agent string.

Client

Client for AWS Single Sign-On

Config

Service config.

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

RetryConfig

Retry configuration for requests.

Enums

Error

All possible error types for this service.

Statics

PKG_VERSION

Crate version number.