Struct aws_sdk_sso::config::interceptors::BeforeTransmitInterceptorContextMut

pub struct BeforeTransmitInterceptorContextMut<'a, I = Input, O = Output, E = Error> { /* private fields */ }

Interceptor context for several hooks in between serialization and transmission.

Only the request is available at this point in the operation.

Implementations

impl<'a, I, O, E> BeforeTransmitInterceptorContextMut<'a, I, O, E>

pub fn request(&self) -> &Request

Returns a reference to the transmittable request for the operation being invoked.

pub fn request_mut(&mut self) -> &mut Request

Returns a mutable reference to the transmittable request for the operation being invoked.

pub fn inner(&self) -> &InterceptorContext<I, O, E>

Downgrade this wrapper struct, returning the underlying InterceptorContext.

There’s no good reason to use this unless you’re writing tests or you have to interact with an API that doesn’t support the context wrapper structs.

pub fn inner_mut(&mut self) -> &mut InterceptorContext<I, O, E>

Downgrade this wrapper struct, returning the underlying InterceptorContext.

There’s no good reason to use this unless you’re writing tests or you have to interact with an API that doesn’t support the context wrapper structs.