aws_sdk_sso::config

Trait IntoShared

pub trait IntoShared<Shared> {
    // Required method
    fn into_shared(self) -> Shared;
}

Like the Into trait, but for (efficiently) converting into a shared type.

If the type is already a shared type, it won’t be nested in another shared type.

See the module docs for information about shared types.

Required Methods

fn into_shared(self) -> Shared

Creates a shared type from an unshared type.