Module aws_sdk_sso::error

Errors that can occur when calling the service.

Modules

invalid_request_exception

See InvalidRequestException

resource_not_found_exception

See ResourceNotFoundException

too_many_requests_exception

See TooManyRequestsException

unauthorized_exception

See UnauthorizedException

Structs

GetRoleCredentialsError

Error type for the GetRoleCredentials operation.

InvalidRequestException

Indicates that a problem occurred with the input to the request. For example, a required parameter might be missing or out of range.

ListAccountRolesError

Error type for the ListAccountRoles operation.

ListAccountsError

Error type for the ListAccounts operation.

LogoutError

Error type for the Logout operation.

ResourceNotFoundException

The specified resource doesn't exist.

TooManyRequestsException

Indicates that the request is being made too frequently and is more than what the server can handle.

UnauthorizedException

Indicates that the request is not authorized. This can happen due to an invalid access token in the request.

Enums

GetRoleCredentialsErrorKind

Types of errors that can occur for the GetRoleCredentials operation.

ListAccountRolesErrorKind

Types of errors that can occur for the ListAccountRoles operation.

ListAccountsErrorKind

Types of errors that can occur for the ListAccounts operation.

LogoutErrorKind

Types of errors that can occur for the Logout operation.