Module aws_sdk_sso::model

Data structures used by operation inputs/outputs.

Modules

account_info

See AccountInfo

role_credentials

See RoleCredentials

role_info

See RoleInfo

Structs

AccountInfo

Provides information about your AWS account.

RoleCredentials

Provides information about the role credentials that are assigned to the user.

RoleInfo

Provides information about the role that is assigned to the user.