Struct aws_sdk_sso::Client

pub struct Client { /* private fields */ }

Client for AWS Single Sign-On

Client for invoking operations on AWS Single Sign-On. Each operation on AWS Single Sign-On is a method on this this struct. .send() MUST be invoked on the generated operations to dispatch the request to the service.

Using the Client

A client has a function for every operation that can be performed by the service. For example, the GetRoleCredentials operation has a Client::get_role_credentials, function which returns a builder for that operation. The fluent builder ultimately has a send() function that returns an async future that returns a result, as illustrated below:

let result = client.get_role_credentials()
    .role_name("example")
    .send()
    .await;

The underlying HTTP requests that get made by this can be modified with the customize_operation function on the fluent builder. See the customize module for more information.

Implementations

impl Client

pub fn get_role_credentials(&self) -> GetRoleCredentialsFluentBuilder

Constructs a fluent builder for the GetRoleCredentials operation.

The fluent builder is configurable:

role_name(impl Into<String>) / set_role_name(Option<String>):
required: true

The friendly name of the role that is assigned to the user.

account_id(impl Into<String>) / set_account_id(Option<String>):
required: true

The identifier for the AWS account that is assigned to the user.

access_token(impl Into<String>) / set_access_token(Option<String>):
required: true

The token issued by the CreateToken API call. For more information, see CreateToken in the IAM Identity Center OIDC API Reference Guide.

On success, responds with GetRoleCredentialsOutput with field(s):

role_credentials(Option<RoleCredentials>):
The credentials for the role that is assigned to the user.

On failure, responds with SdkError<GetRoleCredentialsError>

impl Client

pub fn list_account_roles(&self) -> ListAccountRolesFluentBuilder

Constructs a fluent builder for the ListAccountRoles operation. This operation supports pagination; See into_paginator().

The fluent builder is configurable:

next_token(impl Into<String>) / set_next_token(Option<String>):
required: false

The page token from the previous response output when you request subsequent pages.

max_results(i32) / set_max_results(Option<i32>):
required: false

The number of items that clients can request per page.

access_token(impl Into<String>) / set_access_token(Option<String>):
required: true

The token issued by the CreateToken API call. For more information, see CreateToken in the IAM Identity Center OIDC API Reference Guide.

account_id(impl Into<String>) / set_account_id(Option<String>):
required: true

The identifier for the AWS account that is assigned to the user.

On success, responds with ListAccountRolesOutput with field(s):

next_token(Option<String>):
The page token client that is used to retrieve the list of accounts.

role_list(Option<Vec::<RoleInfo>>):
A paginated response with the list of roles and the next token if more results are available.

On failure, responds with SdkError<ListAccountRolesError>

impl Client

pub fn list_accounts(&self) -> ListAccountsFluentBuilder

Constructs a fluent builder for the ListAccounts operation. This operation supports pagination; See into_paginator().

The fluent builder is configurable:

next_token(impl Into<String>) / set_next_token(Option<String>):
required: false

(Optional) When requesting subsequent pages, this is the page token from the previous response output.

max_results(i32) / set_max_results(Option<i32>):
required: false

This is the number of items clients can request per page.

access_token(impl Into<String>) / set_access_token(Option<String>):
required: true

The token issued by the CreateToken API call. For more information, see CreateToken in the IAM Identity Center OIDC API Reference Guide.

On success, responds with ListAccountsOutput with field(s):

next_token(Option<String>):
The page token client that is used to retrieve the list of accounts.

account_list(Option<Vec::<AccountInfo>>):
A paginated response with the list of account information and the next token if more results are available.

On failure, responds with SdkError<ListAccountsError>

impl Client

pub fn logout(&self) -> LogoutFluentBuilder

Constructs a fluent builder for the Logout operation.

The fluent builder is configurable:

access_token(impl Into<String>) / set_access_token(Option<String>):
required: true

The token issued by the CreateToken API call. For more information, see CreateToken in the IAM Identity Center OIDC API Reference Guide.

On success, responds with LogoutOutput

On failure, responds with SdkError<LogoutError>

impl Client

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.

Panics

This method will panic in the following cases:

Retries or timeouts are enabled without a sleep_impl configured.

Identity caching is enabled without a sleep_impl and time_source configured.

No behavior_version is provided.

The panic message for each of these will have instructions on how to resolve them.

pub fn config(&self) -> &Config

Returns the client’s configuration.

impl Client

pub fn new(sdk_config: &SdkConfig) -> Self

Creates a new client from an SDK Config.

Panics

This method will panic if the sdk_config is missing an async sleep implementation. If you experience this panic, set the sleep_impl on the Config passed into this function to fix it.

This method will panic if the sdk_config is missing an HTTP connector. If you experience this panic, set the http_connector on the Config passed into this function to fix it.

This method will panic if no BehaviorVersion is provided. If you experience this panic, set behavior_version on the Config or enable the behavior-version-latest Cargo feature.