Module aws_sdk_sso::types::error

Error types that AWS Single Sign-On can respond with.

Modules

builders

Builders

Structs

InvalidRequestException

Indicates that a problem occurred with the input to the request. For example, a required parameter might be missing or out of range.

ResourceNotFoundException

The specified resource doesn't exist.

TooManyRequestsException

Indicates that the request is being made too frequently and is more than what the server can handle.

UnauthorizedException

Indicates that the request is not authorized. This can happen due to an invalid access token in the request.