Crate aws_sdk_sso

AWS IAM Identity Center (successor to AWS Single Sign-On) Portal is a web service that makes it easy for you to assign user access to IAM Identity Center resources such as the AWS access portal. Users can get AWS account applications and roles assigned to them and get federated into the application.

This reference guide describes the IAM Identity Center Portal operations that you can call programatically and includes detailed information on data types and errors.

Getting Started

Examples are available for many services and operations, check out the examples folder in GitHub.

The SDK provides one crate per AWS service. You must add Tokio as a dependency within your Rust project to execute asynchronous code. To add aws-sdk-sso to your project, add the following to your Cargo.toml file:

[dependencies]
aws-config = { version = "1.1.7", features = ["behavior-version-latest"] }
aws-sdk-sso = "1.43.0"
tokio = { version = "1", features = ["full"] }

Then in code, a client can be created with the following:

use aws_sdk_sso as sso;

#[::tokio::main]
async fn main() -> Result<(), sso::Error> {
    let config = aws_config::load_from_env().await;
    let client = aws_sdk_sso::Client::new(&config);

    // ... make some calls with the client

    Ok(())
}

See the client documentation for information on what calls can be made, and the inputs and outputs for each of those calls.

Using the SDK

Until the SDK is released, we will be adding information about using the SDK to the Developer Guide. Feel free to suggest additional sections for the guide by opening an issue and describing what you are trying to do.

Getting Help

GitHub discussions - For ideas, RFCs & general questions

GitHub issues - For bug reports & feature requests

Generated Docs (latest version)

Usage examples

Crate Organization

The entry point for most customers will be Client, which exposes one method for each API offered by AWS Single Sign-On. The return value of each of these methods is a “fluent builder”, where the different inputs for that API are added by builder-style function call chaining, followed by calling send() to get a Future that will result in either a successful output or a SdkError.

Some of these API inputs may be structs or enums to provide more complex structured information. These structs and enums live in types. There are some simpler types for representing data such as date times or binary blobs that live in primitives.

All types required to configure a client via the Config struct live in config.

The operation module has a submodule for every API, and in each submodule is the input, output, and error type for that API, as well as builders to construct each of those.

There is a top-level Error type that encompasses all the errors that the client can return. Any other error type can be converted to this Error type via the From trait.

The other modules within this crate are not required for normal usage.

Modules

client

Client for calling AWS Single Sign-On.

config

Configuration for AWS Single Sign-On.

error

Common errors and error handling utilities.

meta

Information about this crate.

operation

All operations that this crate can perform.

primitives

Primitives such as Blob or DateTime used by other types.

types

Data structures used by operation inputs/outputs.

Structs

Client

Client for AWS Single Sign-On

Config

Configuration for a aws_sdk_sso service client.

Enums

Error

All possible error types for this service.