Crate aws_sdk_sso

AWS IAM Identity Center (successor to AWS Single Sign-On) Portal is a web service that makes it easy for you to assign user access to IAM Identity Center resources such as the AWS access portal. Users can get AWS account applications and roles assigned to them and get federated into the application.

Although AWS Single Sign-On was renamed, the sso and identitystore API namespaces will continue to retain their original name for backward compatibility purposes. For more information, see IAM Identity Center rename.

This reference guide describes the IAM Identity Center Portal operations that you can call programatically and includes detailed information on data types and errors.

AWS provides SDKs that consist of libraries and sample code for various programming languages and platforms, such as Java, Ruby, .Net, iOS, or Android. The SDKs provide a convenient way to create programmatic access to IAM Identity Center and other AWS services. For more information about the AWS SDKs, including how to download and install them, see Tools for Amazon Web Services.

Crate Organization

The entry point for most customers will be Client. Client exposes one method for each API offered by the service.

Some APIs require complex or nested arguments. These exist in model.

Lastly, errors that can be returned by the service are contained within error. Error defines a meta error encompassing all possible errors that can be returned by the service.

The other modules within this crate are not required for normal usage.

Modules

client

Client and fluent builders for calling the service.

config

Configuration for the service.

error

All error types that operations can return. Documentation on these types is copied from the model.

input

Input structures for operations. Documentation on these types is copied from the model.

middleware

Base Middleware Stack

model

Data structures used by operation inputs/outputs. Documentation on these types is copied from the model.

operation

All operations that this crate can perform.

output

Output structures for operations. Documentation on these types is copied from the model.

paginator

Paginators for the service

types

Data primitives referenced by other data types.

Structs

AppName

App name that can be configured with an AWS SDK client to become part of the user agent string.

Client

Client for AWS Single Sign-On

Config

Service config.

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

Enums

Error

All possible error types for this service.

Statics

PKG_VERSION

Crate version number.